requirements.txt
requests
ruamel.yaml
google.cloud
